+++ platform/viewer/src/components/ViewportGrid/ViewportGrid.js
@@ -91,6 +91,7 @@ const ViewportGrid = function(props) {
 
   return (
     <div
+      data-cy="viewprt-grid"
       style={{
         display: 'grid',
         gridTemplateRows: `repeat(${numRows}, ${rowSize}%)`,
@@ -106,7 +107,7 @@ const ViewportGrid = function(props) {
 
 ViewportGrid.propTypes = {
   viewportData: PropTypes.array.isRequired,
-  supportsDragAndDrop: PropTypes.bool.isRequired,
+  supportsDrop: PropTypes.bool.isRequired,
   activeViewportIndex: PropTypes.number.isRequired,
   layout: PropTypes.object.isRequired,
   availablePlugins: PropTypes.object.isRequired,
@@ -126,7 +127,7 @@ ViewportGrid.defaultProps = {
     viewports: [{}],
   },
   activeViewportIndex: 0,
-  supportsDragAndDrop: true,
+  supportsDrop: true,
   availablePlugins: {
     DefaultViewport,
   },